coin unit number should starts with 1 rather than 0
This will fix the decreasing credit counter issue on
TAITO Type X2.

+++ Jamma.h
@@ -14,7 +14,7 @@ class Jamma final {
   void Update(bool swap);
   uint8_t GetSw(uint8_t index, bool rapid_mode, bool rapid_mask);
   uint8_t GetCoin(uint8_t index);
-  void SubCoin(int index, uint8_t sub);
+  void SubCoin(uint8_t index, uint8_t sub);
 
  private:
   uint8_t sw[4];
